+++ great_international/panels/great_international.py
@@ -55,6 +55,34 @@ class InternationalHomePagePanels:
                     ]
                 ),
                 MultiFieldPanel(
+                    heading="How DIT helps",
+                    classname='collapsible',
+                    children=[
+                        HelpPanel('Required for section to show: title and at least one related page'),
+                        FieldPanel('how_dit_help_title'),
+                        FieldRowPanel([
+                            MultiFieldPanel([
+                                PageChooserPanel(
+                                    'related_how_dit_help_page_one',
+                                    'great_international.AboutDitServicesPage'
+                                ),
+                            ]),
+                            MultiFieldPanel([
+                                PageChooserPanel(
+                                    'related_how_dit_help_page_two',
+                                    'great_international.AboutDitServicesPage'
+                                ),
+                            ]),
+                            MultiFieldPanel([
+                               PageChooserPanel(
+                                   'related_how_dit_help_page_three',
+                                   'great_international.AboutDitServicesPage'
+                               ),
+                            ]),
+                        ]),
+                    ],
+                ),
+                MultiFieldPanel(
                     heading='Features highlight',
                     children=[
                         FieldPanel('section_two_heading'),