+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - CI-508 - Allowing DIT services pages to be children of invest, capital invest and trade home pages
 - CI-510 - Updated help panel to mirror front-end changes
 - XOT-1124 - Add icon and pages count to industry tags
+- CI-518 - Added new how dit help fields to International Home page
 
 ## [2019.09.19](https://github.com/uktrade/directory-cms/releases/tag/2019.09.19)
 [Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.09.18...2019.09.19)